requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.1            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
django-rest-swagger==2.1.2
django==1.11.2
djangorestframework==3.6.3
idna==2.5                 # via requests
itypes==1.1.0             # via coreapi
jinja2==2.9.6             # via coreschema
markupsafe==1.0           # via jinja2
openapi-codec==1.3.1      # via django-rest-swagger
psycopg2==2.7.1
py==1.4.34                # via pytest
pytest-django==3.1.2
pytest==3.1.2             # via pytest-django
pytz==2017.2              # via django
requests==2.18.1          # via coreapi
simplejson==3.11.1        # via django-rest-swagger
uritemplate==3.0.0        # via coreapi
urllib3==1.21.1           # via requests
